new xml features

+++ src/main/java/com/poc/exceloperations/AppRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AppRepository extends JpaRepository<EmployeeDetails, Long> {
+public interface AppRepository extends JpaRepository<EmployeeDetails, Integer> {
 
 }
